requirements.txt


# Weaviate vector database Python API
weaviate-client==4.11.0

# For loading environmental variables such as API
# keys and endpoint URLs
python-dotenv==1.0.0

# OpenAI Python API for interacting with generative
# and embedding models
openai==1.64.0

anthropic == 0.50.0

firecrawl-py == 2.2.0
